requirements.txt
# Intel NPU optimized libraries
openvino>=2024.0.0
openvino-dev>=2024.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.8.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Text processing
transformers>=4.35.0
tokenizers>=0.14.0

# Utilities
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.1